From: Jan Kára <jack@suse.cz> Subject: Allow direct IO with 512 byte alignment
Newer Linux kernels handle direct IO with alignment of 512b only. So allow
this with dd_rescue as well. This is useful to recover the rest of page-sized
chunk from disk because buffered writes always read multiple of page size and
thus return with EIO when any sector underlying a page is bad.

Signed-off-by: Jan Kara <jack@suse.cz>